rename noise
